h5 read/write ok for basic types
- todo : finish and test for triqs types
