+++ triqs/h5/variant.hpp
@@ -1,3 +1,4 @@
+
 /*******************************************************************************
  *
  * TRIQS: a Toolbox for Research in Interacting Quantum Systems
@@ -38,22 +39,36 @@ namespace std { // to be found by ADL
 namespace triqs::h5 {
 
   /**
-   * Optional : write if the value is set.
    */
   template <typename... T> void h5_write(h5::group gr, std::string const &name, std::variant<T...> const &v) {
     visit([&](auto const &x) { h5_write(gr, name, x); }, v);
   }
 
+  template <typename T> bool h5_is(datatype dt) {
+    // special case for the string
+    if constexpr (std::is_same<T, std::string>::value) { return H5Tget_class(dt) == H5T_STRING; }
+    //if constexpr (std::is_integral<T>::value) { return H5Tget_class(dt) == H5T_INTEGER; }
+    //if constexpr (std::is_same<T, double>::value) { return H5Tget_class(dt) == H5T_FLOAT; }
+    return H5Tequal(dt, h5_type_from_C(T{}));
+  }
+
+  template <typename VT, typename U, typename... T> void h5_read_variant_helper(VT &v, datatype dt, h5::group gr, std::string const &name) {
+    if (h5_is<U>(dt)) {
+      v = VT{triqs::h5::h5_read<U>(gr, name)};
+      return;
+    }
+    if constexpr (sizeof...(T) > 0) h5_read_variant_helper<VT, T...>(v, dt, gr, name);
+  }
+
   /**
    * Read optional from the h5
    */
-  template <typename... T> void h5_read(h5::group gr, std::string name, std::variant<T...> &v) {
-    visit([&](auto &x) { 
-	
-        dataset ds = gr.open_dataset(name);
- 	std::cout  << " eq " << H5Tequal(H5Dget_type(ds), h5_type_from_C(x))<<std::endl;;
-
-	
-	h5_read(gr, name, x); }, v);
+  template <typename... T> void h5_read(h5::group gr, std::string const &name, std::variant<T...> &v) {
+    // name is a group --> triqs object
+    // assume for the moment, name is a dataset.
+    dataset ds  = gr.open_dataset(name);
+    datatype dt = H5Dget_type(ds);
+    h5_read_variant_helper<std::variant<T...>, T...>(v, dt, gr, name);
   }
+
 } // namespace triqs::h5
